first wave of patches to conform to stricter warning policy

+++ Source/Marker/OCRLineMarker.cpp
@@ -28,7 +28,7 @@ OCRLineMarker::~OCRLineMarker() {
 void OCRLineMarker::getOcrLinePoints(const cv::Point& seed,
 									 const cv::Size& imageSize,
 									 std::vector<cv::Point>& points){
-	int height = imageSize.height*imagePercentage_;
+    int height = (int)(imageSize.height*imagePercentage_);
 	int up = seed.y - (height >> 1);
 	int lo = seed.y + (height >> 1);
 	if (up < 0){
